+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/config/LambdaConfig.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.aws.lambda;
+package com.startingblue.fourtooncookie.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,9 +7,6 @@ import software.amazon.awssdk.core.retry.RetryPolicy;
 import software.amazon.awssdk.http.apache.ApacheHttpClient;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.lambda.LambdaClient;
-import software.amazon.awssdk.http.apache.ApacheHttpClient;
-
-import java.time.Duration;
 
 import java.time.Duration;
